chore: inside a JSON or Excel file, resource & property names must be unique (DEV-1971) (#345)
In a long Excel file, it can easily happen that there are two rows with the same property/resource name. If such an Excel is converted to JSON, this mistake will remain unnoticed, because in the JSON schema, it is not possible to make such a check. Only when the data model is created with dsp-tools create, the user is notified.

It is necessary to catch this mistake at an early time, when the Excel is converted to JSON, so that the user is notified about the problematic row in the Excel file.