chore: don't configure logging statically (DEV-2183) (#385)
Currently, logging is configured statically with `logging.basicConfig()`. But this has been shown to interfere with users of dsp-tools who try to configure logging statically.

For example, when users of the import script (https://github.com/dasch-swiss/00A1-import-scripts/blob/main/import_script.py) want to log to a file, they want the easiest configuration possible, which is the static `logging.basicConfig()`. But this fails, because at the top of their script, they import excel2xml from DSP-TOOLS, so logging is already configured statically. 

In libraries, it is generally discouraged to log to the root logger: https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library