chore(xmlupload): improve error handling of Timeout Error (DEV-2299) (#411)
Old behaviour in case of a timeout:

```
dsp-tools xmlupload testdata/xml-data/test-data-systematic.xml
...
2023-06-19T17:26:59.714574: Try reconnecting to DSP server, next attempt in 1 seconds...
2023-06-19T17:27:00.896490: Try reconnecting to DSP server, next attempt in 2 seconds...
2023-06-19T17:27:03.032554: Try reconnecting to DSP server, next attempt in 4 seconds...
2023-06-19T17:27:07.158967: Try reconnecting to DSP server, next attempt in 8 seconds...
2023-06-19T17:27:15.297954: Try reconnecting to DSP server, next attempt in 16 seconds...
2023-06-19T17:27:31.423494: Try reconnecting to DSP server, next attempt in 32 seconds...
2023-06-19T17:28:03.569195: Try reconnecting to DSP server, next attempt in 64 seconds...
WARNING: Unable to upload file 'testdata/bitstreams/test.mp4' of resource 'Moviething' (video_thing_1): Permanently unable to execute the network action. See logs for more details....
...
WARNING: Unable to create resource 'First Testthing' (test_thing_1): Target of link is not a IRI: "video_thing_1"
...
Upload the stashed resptrs...
  Upload resptrs of resource "test_thing_0"...
    WARNING: Unable to upload the resptr prop of 'testonto:hasTestThing' of resource 'test_thing_0'. Original error message: Invalid IRI: test_thing_1

==========================================
xmlupload must be aborted because of an error
The mapping of internal IDs to IRIs was written to ...
There are stashed text properties that could not be reapplied to the resources they were stripped from. They were saved to ...
There are stashed resptr properties that could not be reapplied to the resources they were stripped from. They were saved to ...
Independently of this error, there were some resources that could not be uploaded: ...
==========================================

Traceback (most recent call last):
... (LONG STACKTRACE)
```

New behaviour in case of a timeout:

```
xmlupload testdata/xml-data/test-data-systematic.xml
...
2023-06-20T10:41:27.924901: Timeout Error: Try reconnecting to DSP server, next attempt in 1 seconds...
2023-06-20T10:41:29.258539: Timeout Error: Try reconnecting to DSP server, next attempt in 2 seconds...
2023-06-20T10:41:31.577019: Timeout Error: Try reconnecting to DSP server, next attempt in 4 seconds...
2023-06-20T10:41:35.909626: Timeout Error: Try reconnecting to DSP server, next attempt in 8 seconds...
2023-06-20T10:41:44.252310: Timeout Error: Try reconnecting to DSP server, next attempt in 16 seconds...
2023-06-20T10:42:00.574545: Timeout Error: Try reconnecting to DSP server, next attempt in 32 seconds...
2023-06-20T10:42:32.899971: Timeout Error: Try reconnecting to DSP server, next attempt in 64 seconds...
WARNING: Unable to upload file 'testdata/bitstreams/test.mp4' of resource 'Moviething' (video_thing_1): Permanently unable to execute the network action. See logs for more details.
...
WARNING: Unable to create resource 'First Testthing' (test_thing_1): Target of link is not a IRI: "video_thing_1"
...
Upload the stashed resptrs...
  Upload resptrs of resource "test_thing_0"...
    WARNING: Unable to upload the resptr prop of 'testonto:hasTestThing' of resource 'test_thing_0'. Original error message: Invalid IRI: test_thing_1

==========================================
xmlupload must be aborted because of an error.
Error message: 'Some stashed resptrs or XML texts could not be reapplied to their resources on the DSP server.'

The mapping of internal IDs to IRIs was written to ...

There are stashed text properties that could not be reapplied to the resources they were stripped from. They were saved to ...

There are stashed resptr properties that could not be reapplied to the resources they were stripped from. They were saved to ...

Independently of this error, there were some resources that could not be uploaded: ...
```