fix(xmlupload): make sure resources aren't created multiple times after timeout error (DEV-2412) #432
If the network connection is very slow, it can happen that the client raises a timeout error, while the API is still processing the request. Because the client has only a temporary identifier of the resource, but no IRI, it is possible to create the same resource multiple times.

The solution is to set the timeout to None, and to wait until the response is available, even if this takes very long.