simplify test excel2resources
Before:
The function created the resource section and in the class set up saved it as a json in a temp folder. It then read the json and checked that it was the same as the input from the method.
Saving the json is not something we need to test.
But it significantly complicated this test.
Instead now we just take the output from the function and extract the information. No temp needed.